메일 자동 작성/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
PyMySQL==1.1.0
SQLAlchemy>=2.0.0